fixed overflow and reduced test size

+++ lab3/insertionSort.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 #include <vector>
 #include "insertionSort.hpp"
-static int insert_compare=0;
+static unsigned long long int insert_compare=0;
 void InsertionSort( std::vector<int> &a) {
     int i, j, key, numLength = a.size();
     for (j = 1; j < numLength; j++)   // Start with 1 (not 0)
@@ -15,8 +15,8 @@ void InsertionSort( std::vector<int> &a) {
         a[i + 1] = key;  //Put key into its proper location
     }
 }
-int getInsertCompare() {
-    int i=insert_compare;
+unsigned long long int getInsertCompare() {
+    unsigned long long int i=insert_compare;
     insert_compare = 0;
     return i;
 }
